refactor: granulates storage my offers page with atoms and molecules
