+++ src/components/pages/storage/index.ts
@@ -1,7 +1,7 @@
 import StorageSellPage from './sell/StorageSellPage'
 import StorageOffersPage from './buy/StorageOffersPage'
 import StorageLandingPage from './StorageLandingPage'
-import StorageMyOffersPage from './sell/StorageMyOffersPage'
+import StorageMyOffersPage from './myoffers/StorageMyOffersPage'
 
 export {
   StorageSellPage,
